Saves profiler snapshots every 5 minutes (#83467)
As the title says. Each saved snapshot will be named
- `profiler-[TIME IN SECONDS].json`
- `sendmaps-[TIME IN SECONDS].json`

For example, `profiler-0.json`, `profiler-300.json`, `profiler-600.json`
would correspond to a profile at the start of the round, a profile 300
seconds into a round and a profile 600 seconds into a round.

The timings depend on world.time, so it's probably not a good idea to
rely on these timings to stay consistent.

Allows us to track performance of procs over time. Could allow us to
spot any sort of anomalies or performance sinks.

CC: @MrStonedOne @bobbah @LemonInTheDark

---------

Co-authored-by: Watermelon914 <3052169-Watermelon914@users.noreply.gitlab.com>